Separate all event-handling to smaller modules
