+++ spring-cloud-alibaba-dubbo/src/main/java/org/springframework/cloud/alibaba/dubbo/autoconfigure/DubboMetadataAutoConfiguration.java
@@ -17,21 +17,29 @@
 package org.springframework.cloud.alibaba.dubbo.autoconfigure;
 
 import org.apache.dubbo.config.ProtocolConfig;
+import org.apache.dubbo.config.spring.ServiceBean;
+import org.apache.dubbo.config.spring.context.event.ServiceBeanExportedEvent;
 
 import feign.Contract;
 import org.springframework.beans.factory.ObjectProvider;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.context.event.ApplicationFailedEvent;
+import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.cloud.alibaba.dubbo.metadata.DubboProtocolConfigSupplier;
 import org.springframework.cloud.alibaba.dubbo.metadata.repository.DubboServiceMetadataRepository;
 import org.springframework.cloud.alibaba.dubbo.metadata.resolver.DubboServiceBeanMetadataResolver;
 import org.springframework.cloud.alibaba.dubbo.metadata.resolver.MetadataResolver;
 import org.springframework.cloud.alibaba.dubbo.service.DubboGenericServiceFactory;
+import org.springframework.cloud.alibaba.dubbo.service.DubboMetadataConfigServiceExporter;
 import org.springframework.cloud.alibaba.dubbo.service.DubboMetadataConfigServiceProxy;
 import org.springframework.cloud.alibaba.dubbo.service.PublishingDubboMetadataConfigService;
 import org.springframework.cloud.alibaba.dubbo.util.JSONUtils;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
+import org.springframework.context.event.ContextClosedEvent;
+import org.springframework.context.event.EventListener;
 
 import java.util.Collection;
 import java.util.function.Supplier;
@@ -42,8 +50,20 @@ import java.util.function.Supplier;
  * @author <a href="mailto:mercyblitz@gmail.com">Mercy</a>
  */
 @Configuration
-@Import({DubboServiceMetadataRepository.class, PublishingDubboMetadataConfigService.class, JSONUtils.class})
+@Import({DubboServiceMetadataRepository.class,
+        PublishingDubboMetadataConfigService.class,
+        DubboMetadataConfigServiceExporter.class,
+        JSONUtils.class})
 public class DubboMetadataAutoConfiguration {
+
+    @Autowired
+    private PublishingDubboMetadataConfigService dubboMetadataConfigService;
+
+    @Autowired
+    private MetadataResolver metadataResolver;
+
+    @Autowired
+    private DubboMetadataConfigServiceExporter dubboMetadataConfigServiceExporter;
 
     @Bean
     @ConditionalOnMissingBean
@@ -60,5 +80,32 @@ public class DubboMetadataAutoConfiguration {
     @ConditionalOnMissingBean
     public DubboMetadataConfigServiceProxy dubboMetadataConfigServiceProxy(DubboGenericServiceFactory factory) {
         return new DubboMetadataConfigServiceProxy(factory);
+    }
+
+    // Event-Handling
+
+    @EventListener(ServiceBeanExportedEvent.class)
+    public void onServiceBeanExported(ServiceBeanExportedEvent event) {
+        ServiceBean serviceBean = event.getServiceBean();
+        publishServiceRestMetadata(serviceBean);
+    }
+
+    private void publishServiceRestMetadata(ServiceBean serviceBean) {
+        dubboMetadataConfigService.publishServiceRestMetadata(metadataResolver.resolveServiceRestMetadata(serviceBean));
+    }
+
+    @EventListener(ApplicationReadyEvent.class)
+    public void onApplicationReady() {
+        dubboMetadataConfigServiceExporter.export();
+    }
+
+    @EventListener(ApplicationFailedEvent.class)
+    public void onApplicationFailed() {
+        dubboMetadataConfigServiceExporter.unexport();
+    }
+
+    @EventListener(ContextClosedEvent.class)
+    public void onContextClosed() {
+        dubboMetadataConfigServiceExporter.unexport();
     }
 }
