+++ spring-cloud-alibaba-dubbo/src/main/java/org/springframework/cloud/alibaba/dubbo/autoconfigure/DubboServiceRegistrationAutoConfiguration.java
@@ -17,35 +17,27 @@
 package org.springframework.cloud.alibaba.dubbo.autoconfigure;
 
 import org.apache.dubbo.common.URL;
-import org.apache.dubbo.config.ApplicationConfig;
-import org.apache.dubbo.config.ProtocolConfig;
-import org.apache.dubbo.config.ServiceConfig;
-import org.apache.dubbo.config.spring.ServiceBean;
 import org.apache.dubbo.config.spring.context.event.ServiceBeanExportedEvent;
 
 import com.ecwid.consul.v1.agent.model.NewService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.context.event.ApplicationFailedEvent;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.cloud.alibaba.dubbo.metadata.repository.DubboServiceMetadataRepository;
-import org.springframework.cloud.alibaba.dubbo.metadata.resolver.MetadataResolver;
+import org.springframework.cloud.alibaba.dubbo.registry.DubboServiceRegistrationEventPublishingAspect;
 import org.springframework.cloud.alibaba.dubbo.registry.event.ServiceInstancePreRegisteredEvent;
-import org.springframework.cloud.alibaba.dubbo.service.DubboMetadataConfigService;
-import org.springframework.cloud.alibaba.dubbo.service.PublishingDubboMetadataConfigService;
+import org.springframework.cloud.alibaba.dubbo.service.DubboMetadataConfigServiceExporter;
 import org.springframework.cloud.alibaba.dubbo.util.JSONUtils;
 import org.springframework.cloud.client.ServiceInstance;
 import org.springframework.cloud.client.serviceregistry.Registration;
 import org.springframework.cloud.client.serviceregistry.ServiceRegistry;
 import org.springframework.cloud.consul.serviceregistry.ConsulRegistration;
 import org.springframework.context.ApplicationListener;
-import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.event.ContextClosedEvent;
+import org.springframework.context.annotation.Import;
 import org.springframework.context.event.EventListener;
 import org.springframework.util.CollectionUtils;
 import org.springframework.util.StringUtils;
@@ -53,19 +45,19 @@ import org.springframework.util.StringUtils;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
-import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
 import static org.springframework.cloud.alibaba.dubbo.registry.SpringCloudRegistry.DUBBO_URLS_METADATA_PROPERTY_NAME;
 
 /**
- * The Auto-Configuration class for Dubbo metadata {@link EventListener event handling}.
+ * Dubbo Service Registration Auto-{@link Configuration}
  *
  * @author <a href="mailto:mercyblitz@gmail.com">Mercy</a>
  */
-@AutoConfigureAfter(value = {DubboMetadataAutoConfiguration.class})
 @Configuration
-public class DubboMetadataEventHandlingAutoConfiguration {
+@Import({DubboServiceRegistrationEventPublishingAspect.class})
+@ConditionalOnProperty(value = "spring.cloud.service-registry.auto-registration.enabled", matchIfMissing = true)
+public class DubboServiceRegistrationAutoConfiguration {
 
     private static final String CONSUL_REGISTRATION_CLASS_NAME =
             "org.springframework.cloud.consul.serviceregistry.ConsulAutoRegistration";
@@ -76,44 +68,18 @@ public class DubboMetadataEventHandlingAutoConfiguration {
     private final Logger logger = LoggerFactory.getLogger(getClass());
 
     @Autowired
-    private MetadataResolver metadataResolver;
-
-    @Autowired
-    private PublishingDubboMetadataConfigService dubboMetadataConfigService;
-
-    @Autowired
-    private ApplicationConfig applicationConfig;
-
-    @Autowired
-    private Supplier<ProtocolConfig> protocolConfigSupplier;
-
-    @Autowired
-    private ConfigurableApplicationContext context;
+    private ServiceRegistry serviceRegistry;
 
     @Autowired
     private DubboServiceMetadataRepository dubboServiceMetadataRepository;
 
     @Autowired
-    private JSONUtils jsonUtils;
-
-    @Value("${spring.application.name:application}")
-    private String currentApplicationName;
+    private DubboMetadataConfigServiceExporter dubboMetadataConfigServiceExporter;
 
     @Autowired
-    private ServiceRegistry serviceRegistry;
+    private JSONUtils jsonUtils;
 
-    private volatile Registration registration;
-
-    /**
-     * The ServiceConfig of DubboMetadataConfigService to be exported, can be nullable.
-     */
-    private ServiceConfig<DubboMetadataConfigService> serviceConfig;
-
-    @EventListener(ServiceBeanExportedEvent.class)
-    public void onServiceBeanExported(ServiceBeanExportedEvent event) {
-        ServiceBean serviceBean = event.getServiceBean();
-        publishServiceRestMetadata(serviceBean);
-    }
+    private Registration registration;
 
     @ConditionalOnClass(name = EUREKA_REGISTRATION_CLASS_NAME)
     @Bean
@@ -131,20 +97,15 @@ public class DubboMetadataEventHandlingAutoConfiguration {
         serviceRegistry.register(registration);
     }
 
-    @EventListener(ApplicationFailedEvent.class)
-    public void onApplicationFailed() {
-        unexportDubboMetadataConfigService();
-    }
 
-    @EventListener(ContextClosedEvent.class)
-    public void onContextClosed() {
-        unexportDubboMetadataConfigService();
-    }
+    // Event-Handling
 
     @EventListener(ServiceInstancePreRegisteredEvent.class)
     public void onServiceInstancePreRegistered(ServiceInstancePreRegisteredEvent event) {
+
+        dubboMetadataConfigServiceExporter.export();
+
         Registration registration = event.getSource();
-        exportDubboMetadataConfigService();
         attachURLsIntoMetadata(registration);
         setRegistration(registration);
     }
@@ -195,53 +156,5 @@ public class DubboMetadataEventHandlingAutoConfiguration {
             return null;
         }
         return jsonUtils.toJSON(urls.stream().map(URL::toFullString).collect(Collectors.toList()));
-    }
-
-    private void publishServiceRestMetadata(ServiceBean serviceBean) {
-        dubboMetadataConfigService.publishServiceRestMetadata(metadataResolver.resolveServiceRestMetadata(serviceBean));
-    }
-
-    private void exportDubboMetadataConfigService() {
-
-        if (serviceConfig != null && serviceConfig.isExported()) {
-            return;
-        }
-
-        if (StringUtils.isEmpty(dubboMetadataConfigService.getServiceRestMetadata())) {
-            // If there is no REST metadata, DubboMetadataConfigService will not be exported.
-            if (logger.isInfoEnabled()) {
-                logger.info("There is no REST metadata, the Dubbo service[{}] will not be exported.",
-                        dubboMetadataConfigService.getClass().getName());
-            }
-            return;
-        }
-
-        serviceConfig = new ServiceConfig<>();
-
-        serviceConfig.setInterface(DubboMetadataConfigService.class);
-        // Use current Spring application name as the Dubbo Service version
-        serviceConfig.setVersion(currentApplicationName);
-        serviceConfig.setRef(dubboMetadataConfigService);
-        serviceConfig.setApplication(applicationConfig);
-        serviceConfig.setProtocol(protocolConfigSupplier.get());
-
-        serviceConfig.export();
-
-        if (logger.isInfoEnabled()) {
-            logger.info("The Dubbo service[{}] has been exported.", serviceConfig.toString());
-        }
-    }
-
-    private void unexportDubboMetadataConfigService() {
-
-        if (serviceConfig == null || serviceConfig.isUnexported()) {
-            return;
-        }
-
-        serviceConfig.unexport();
-
-        if (logger.isInfoEnabled()) {
-            logger.info("The Dubbo service[{}] has been unexported.", serviceConfig.toString());
-        }
     }
 }
